+++ ui-admin/src/api/api.tsx
@@ -366,6 +366,7 @@ export type KitRequestListResponse = {
 
 export type InternalConfig = {
   pepperDsmConfig: Record<string, string>
+  addressValidationConfig: Record<string, string>
 }
 
 export type ParticipantTaskUpdateDto = {
